Move V1Beta3Helper classes to V1Beta3
